CCD-2935 Refactor CreateCaseEventDetails model
   [CCD-2935](https://tools.hmcts.net/jira/browse/CCD-2935)

Refactor CreateCaseEventDetails model.
