+++ src/main/java/uk/gov/hmcts/ccd/domain/service/createevent/CreateCaseEventDetails.java
@@ -1,7 +1,9 @@
-package uk.gov.hmcts.ccd.domain.model.definition;
+package uk.gov.hmcts.ccd.domain.service.createevent;
 
 import lombok.Data;
 import uk.gov.hmcts.ccd.domain.model.casedataaccesscontrol.AccessProfile;
+import uk.gov.hmcts.ccd.domain.model.definition.CaseDetails;
+import uk.gov.hmcts.ccd.domain.model.definition.CaseTypeDefinition;
 
 import java.util.Set;
